Anjali/code-analyzer-service/requirements.txt
fastapi
"uvicorn[standard]"
pydantic
python-multipart
google-genai
python-dotenv
uvicorn